ヘルパーメソッドを別ファイルに切り出しました
